Added map deletion Fixed LocalMapRepositoryImpl.kt index

+++ mobile-app/composeApp/src/commonMain/kotlin/data/filesystemProviders/IoFileServiceImpl.kt
@@ -44,7 +44,9 @@ class IoFileServiceImpl(
     override suspend fun deleteFile(directory: String, fileName: String) {
         withContext(Dispatchers.IO) {
             val path = getFullPath(directory, fileName)
-            fileSystem.delete(path)
+            if (fileSystem.exists(path)) {
+                fileSystem.delete(path)
+            }
         }
     }
